requirements.txt
django>=3.2,<5
djangorestframework>=3.14.0
djangorestframework-simplejwt
psycopg2-binary>=2.9
django-cors-headers
gunicorn>=21.2.0
Flask
numpy<2.0,>=1.24.0
tensorflow==2.8.0
psutil
Pillow
deepface>=0.0.75
opencv-python-headless>=4.5.5.62
tf-keras
protobuf<=3.20.3
mtcnn
